implement db clear and load save point, fix auth user data loading
Worked out how to merge the user model saved to the DB, and the user received from the auth listener. 

When I clear the DB, it deletes the saved user but not the user's firebase credentials. So they are still signed in, but the absence of the merged data from the db broke my app.
Now, if there is no database user data, I assign some default values to the auth user state.